# feat: implement GroupsAccumulator for `count(DISTINCT)` aggr