--- a/datafusion/sqllogictest/test_files/aggregate.slt
+++ b/datafusion/sqllogictest/test_files/aggregate.slt
@@ -2302,6 +2302,12 @@ SELECT count(c1, c2) FROM test
 query error DataFusion error: This feature is not implemented: COUNT DISTINCT with multiple arguments
 SELECT count(distinct c1, c2) FROM test
 
+# count(distinct) and count() together
+query II
+SELECT count(c1), count(distinct c1) FROM test
+----
+4 3
+
 # count_null
 query III
 SELECT count(null), count(null, null), count(distinct null) FROM test