@Dandandan 2025-03-20
It would be worthwhile to run the `clickbench_extended` benchmarks as well (`./bench.sh run clickbench_extended`)