@alamb 2025-03-21
> It would be worthwhile to run the `clickbench_extended` benchmarks as well (`./bench.sh run clickbench_extended`)

I am running the benchmarks now and will report