@waynexia 2025-03-21
Ahh, I reproduced the same result. And I also observed a regression on q0:

| Query | Before (ms) | After (ms) |
|-------|-------------|------------|
| Q0    | 1407.04     | 2013.42    |
| Q2    | 681.78      | 1742.08    |

(BTW, how do you get the comparison output