@alamb 2025-03-21
> (BTW, how do you get the comparison output

I am using [`bench.sh` ](https://github.com/apache/datafusion/blob/main/benchmarks/bench.sh)with `bench.sh compare`