@waynexia 2025-03-25
Quick update: made some progress, but still need a few days to refine it. I've implemented a primitive aggregator and it does work